Move ShippingAddressElement to elements package.
