+++ paymentsheet/src/main/java/com/stripe/android/checkout/CheckoutPresenter.kt
@@ -3,6 +3,7 @@ package com.stripe.android.checkout
 import androidx.annotation.RestrictTo
 import com.stripe.android.elements.CurrencySelectorElement
 import com.stripe.android.elements.PaymentElement
+import com.stripe.android.elements.ShippingAddressElement
 import com.stripe.android.paymentelement.CheckoutSessionPreview
 import javax.inject.Inject
 import javax.inject.Provider
